rename WalletIslandDefault to WalletIslandDraggable
