+++ playground/nextjs-app-router/components/form/active-component.tsx
@@ -47,8 +47,8 @@ export function ActiveComponent() {
           <SelectItem value={OnchainKitComponent.WalletDefault}>
             WalletDefault
           </SelectItem>
-          <SelectItem value={OnchainKitComponent.WalletIslandDefault}>
-            WalletIslandDefault
+          <SelectItem value={OnchainKitComponent.WalletIslandDraggable}>
+            WalletIslandDraggable
           </SelectItem>
           <SelectItem value={OnchainKitComponent.NFTCard}>NFT Card</SelectItem>
           <SelectItem value={OnchainKitComponent.NFTCardDefault}>
