+++ playground/nextjs-app-router/types/onchainkit.ts
@@ -10,7 +10,7 @@ export enum OnchainKitComponent {
   TransactionDefault = 'transaction-default',
   Wallet = 'wallet',
   WalletDefault = 'wallet-default',
-  WalletIslandDefault = 'wallet-island-default',
+  WalletIslandDraggable = 'wallet-island-draggable',
   NFTCard = 'nft-card',
   NFTCardDefault = 'nft-card-default',
   NFTMintCard = 'nft-mint-card',
